Break mat2x3Functions into separate files
