+++ src/internal/dataClasses.ts
@@ -12,10 +12,6 @@ class Line implements ILine {
   constructor(public x0 = NaN, public y0 = NaN, public dirX = NaN, public dirY = NaN) {}
 }
 
-class Mat2x3 implements IMat2x3 {
-  constructor(public a = NaN, public b = NaN, public c = NaN, public d = NaN, public e = NaN, public f = NaN) {}
-}
-
 class Ray implements IRay {
   constructor(public x0 = NaN, public y0 = NaN, public dirX = NaN, public dirY = NaN) {}
 }
@@ -38,10 +34,6 @@ export function _intersectionAlloc(): IIntersection {
 
 export function _lineAlloc(): ILine {
   return new Line();
-}
-
-export function _mat2x3Alloc(): IMat2x3 {
-  return new Mat2x3();
 }
 
 export function _polygonAlloc(): IPolygon {
