+++ src/mat2x3Functions/mat2x3AffDeterminant.ts
@@ -1,0 +1,5 @@
+import { IMat2x3 } from "../types";
+
+export function mat2x3AffDeterminant(mat: IMat2x3) {
+  return mat.a * mat.d - mat.b * mat.c;
+}
